requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
multidict==6.0.2
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==2.70.0
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
yarl==1.7.2
zope.interface==5.4.0
